fixed logger circlejerk
